requirements.txt
langchain
langchain-community
langchain-openai
langgraph
langsmith
chromadb
openai
pypdf
python-docx
streamlit
faiss-cpu
